refactor(finance_agent): one loop for both finance agent benchmarks
finance_agent_v2 was a fork of the finance_agent loop that differed only in
policy: the nudge sent on a prose-only turn, a wall-clock budget instead of a
turn cap, and aborting on retry exhaustion. Fold those into the one loop as
required config fields so each server instance states its own policy, and drop
the fork.

The upstream parity checks move to resources_servers/finance_agent_v2/tests,
where the vals-ai package is installed, and now compare the committed config
values against it rather than constants in the agent.
