+++ responses_api_agents/finance_agent/app.py
@@ -12,10 +12,26 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""Multi-turn tool-calling loop for the Vals AI finance agent benchmarks.
+
+One loop serves both benchmarks. Everything that differs between them is a
+config value, not a branch in this file: the nudge injected on a prose-only
+turn, whether the run is bounded by turns or by wall clock, and which tool
+errors abort a rollout instead of being handed back to the model. Those three
+fields are required, so a config states its policy rather than inheriting a
+default that happens to suit the other benchmark.
+
+The values each benchmark uses are pinned against the upstream package by tests
+in ``resources_servers/finance_agent_v2/tests``, which is where that package is
+installed.
+"""
+
 import asyncio
 import json
 import logging
 import re
+import time
+from enum import Enum
 from typing import Any, List, Optional
 
 from fastapi import Request, Response
@@ -50,8 +66,8 @@ logger = logging.getLogger(__name__)
 
 _MODEL_OUTPUT_TYPES = frozenset({"reasoning", "function_call"})
 
-# Regex that matches common vLLM / OpenAI context-length error messages.
-# Mirrors the detection used by the upstream finance-agent benchmark.
+# The model is reached over HTTP, so a context-window rejection arrives as error
+# text rather than a typed exception and has to be recognised by shape.
 _CONTEXT_OVERFLOW_RE = re.compile(
     r"maximum context length is \d+ tokens|"
     r"context length is (?:only )?\d+ tokens|"
@@ -77,12 +93,56 @@ _CONTEXT_OVERFLOW_RE = re.compile(
 )
 
 
+def _decoded_object(raw: str) -> Optional[dict]:
+    """JSON-decode *raw*, returning None unless it decodes to an object."""
+    try:
+        value = json.loads(raw)
+    except (json.JSONDecodeError, TypeError):
+        return None
+    return value if isinstance(value, dict) else None
+
+
+class StopReason(str, Enum):
+    """Why the loop ended.
+
+    Recorded on the rollout so a zero from "never submitted" is distinguishable
+    from a zero from "judged wrong".
+    """
+
+    DONE_TOOL = "done_tool"
+    MAX_TURNS = "max_turns"
+    MAX_TIME = "max_time"
+    MAX_OUTPUT_TOKENS = "max_output_tokens"
+    ERROR = "error"
+
+
 class FinanceAgentConfig(BaseResponsesAPIAgentConfig):
     resources_server: ResourcesServerRef
     model_server: ModelServerRef
-    max_steps: Optional[int] = None
+    no_tool_call_nudge: str = Field(
+        ...,
+        description="Injected as a user message when a turn produces prose and "
+        "no tool call, so the loop continues instead of stopping.",
+    )
+    max_time_seconds: Optional[float] = Field(
+        ...,
+        description="Wall-clock budget for the loop, checked before each model "
+        "call, or null for no budget. This is the agent's hard stop; a resource "
+        "server's max_rollout_time_seconds is a softer budget that only makes "
+        "tools return an error asking the model to submit.",
+    )
+    abort_on_tool_error_types: List[str] = Field(
+        ...,
+        description="Exception type names in a tool's error payload that abort "
+        "the whole rollout instead of being fed back to the model. Empty means "
+        "every tool error is fed back and the run continues.",
+    )
+    max_steps: Optional[int] = Field(
+        default=None,
+        description="Maximum model turns, or null for no turn cap.",
+    )
     done_tools: List[str] = Field(
-        default=["submit_final_result"],
+        default_factory=lambda: ["submit_final_result"],
         description="Tool names that signal the agent loop should terminate. "
         "When any tool call in a batch matches, remaining calls are skipped "
         "and the loop exits.",
@@ -156,6 +216,44 @@ class FinanceAgent(SimpleResponsesAPIAgent):
 
         return outputs[i:]
 
+    @staticmethod
+    def _tool_error_message(tool_output: str) -> Optional[str]:
+        """Return the ``error`` string a tool reply carries, if any.
+
+        Two shapes arrive here. A reply from the resource server is wrapped as
+        ``{"results": "<payload>"}`` where the payload is itself a JSON string,
+        so an error is double-encoded and the inner object must be decoded
+        separately. A tool call that never reached the server (timeout,
+        transport failure) is rendered locally as a flat ``{"error": ...}``.
+
+        Only an ``error`` field counts, so a successful reply whose text merely
+        mentions an exception name is not mistaken for a failure.
+        """
+        payload = _decoded_object(tool_output)
+        if payload is None:
+            return None
+
+        results = payload.get("results")
+        if isinstance(results, str):
+            inner = _decoded_object(results)
+            if inner is not None and isinstance(inner.get("error"), str):
+                return inner["error"]
+
+        error = payload.get("error")
+        return error if isinstance(error, str) else None
+
+    def _aborting_error_type(self, tool_output: str) -> Optional[str]:
+        """Return the abort-worthy exception type named in *tool_output*, if any."""
+        if not self.config.abort_on_tool_error_types:
+            return None
+        error = self._tool_error_message(tool_output)
+        if error is None:
+            return None
+        return next(
+            (name for name in self.config.abort_on_tool_error_types if error.startswith(f"{name}:")),
+            None,
+        )
+
     async def responses(
         self,
         request: Request,
@@ -176,9 +274,26 @@ class FinanceAgent(SimpleResponsesAPIAgent):
 
         done_tools_set = set(self.config.done_tools)
         max_steps = self.config.max_steps
+        max_time_seconds = self.config.max_time_seconds
+        started_at = time.monotonic()
+        stop_reason = StopReason.ERROR
 
         # Check max_steps at the TOP so we never start a turn past the limit.
         while max_steps is None or step < max_steps:
+            # Checked before the turn's query, so an exhausted budget ends the
+            # run rather than paying for one more call.
+            if max_time_seconds is not None:
+                elapsed = time.monotonic() - started_at
+                if elapsed >= max_time_seconds:
+                    logger.warning(
+                        "Time budget exhausted (%.0fs / %.0fs) after %d steps — terminating agent loop",
+                        elapsed,
+                        max_time_seconds,
+                        step,
+                    )
+                    stop_reason = StopReason.MAX_TIME
+                    break
+
             step += 1
             new_body = body.model_copy(update={"input": body.input + new_outputs})
 
@@ -201,6 +316,7 @@ class FinanceAgent(SimpleResponsesAPIAgent):
                     self.config.model_call_timeout,
                     step,
                 )
+                stop_reason = StopReason.ERROR
                 break
             except Exception as e:
                 if self.config.truncate_on_overflow and self._is_context_overflow_error(e):
@@ -215,6 +331,7 @@ class FinanceAgent(SimpleResponsesAPIAgent):
                         new_outputs = truncated
                         continue
                 logger.error("Model call failed on step %d: %s: %s", step, type(e).__name__, e)
+                stop_reason = StopReason.ERROR
                 break
 
             output = model_response.output
@@ -224,6 +341,7 @@ class FinanceAgent(SimpleResponsesAPIAgent):
             usage = accumulate_response_usage(usage, model_response.usage)
 
             if model_response.incomplete_details and model_response.incomplete_details.reason == "max_output_tokens":
+                stop_reason = StopReason.MAX_OUTPUT_TOKENS
                 break
 
             all_fn_calls: List[NeMoGymResponseFunctionToolCall] = [o for o in output if o.type == "function_call"]
@@ -232,13 +350,13 @@ class FinanceAgent(SimpleResponsesAPIAgent):
             ]
 
             if not all_fn_calls and all_output_messages:
-                # Match vals-ai/finance-agent eval (get_agent.py _before_query +
-                # _should_stop=False): inject "Continue." instead of breaking so
-                # the model keeps looping until submit_final_result or max_steps.
-                new_outputs.append(NeMoGymEasyInputMessage(role="user", content="Continue."))
+                # Prose between tool calls does not end the run; the nudge keeps
+                # the model going until it calls a done tool or a limit is hit.
+                new_outputs.append(NeMoGymEasyInputMessage(role="user", content=self.config.no_tool_call_nudge))
                 continue
 
             done = False
+            aborted = False
             for output_function_call in all_fn_calls:
                 try:
                     coro = self.server_client.post(
@@ -289,13 +407,29 @@ class FinanceAgent(SimpleResponsesAPIAgent):
                         output_function_call.name,
                     )
                     done = True
+                    stop_reason = StopReason.DONE_TOOL
                     break
 
-            if done:
-                break
+                if error_type := self._aborting_error_type(tool_output):
+                    logger.error(
+                        "Tool '%s' raised %s — aborting rollout",
+                        output_function_call.name,
+                        error_type,
+                    )
+                    aborted = True
+                    stop_reason = StopReason.ERROR
+                    break
 
-        if max_steps is not None and step >= max_steps:
+            if done or aborted:
+                break
+        else:
+            if max_steps is not None:
+                stop_reason = StopReason.MAX_TURNS
+
+        if stop_reason is StopReason.MAX_TURNS:
             logger.warning("Reached max_steps=%d — terminating agent loop", max_steps)
+
+        logger.info("Agent loop finished after %d steps: stop_reason=%s", step, stop_reason.value)
 
         if last_model_response is None:
             logger.error("Agent loop terminated without any successful model response")
@@ -318,6 +452,16 @@ class FinanceAgent(SimpleResponsesAPIAgent):
 
         last_model_response.output = new_outputs
         last_model_response.usage = usage
+        # Carry why the loop ended into the rollout. Without this the stop reason
+        # exists only in the agent's log, so a results file cannot distinguish an
+        # answer the model chose to submit from one cut short by the time budget or
+        # a tool abort. ``metadata`` is the Responses API's own string-keyed side
+        # channel, so it survives verify() untouched.
+        last_model_response.metadata = {
+            **(last_model_response.metadata or {}),
+            "stop_reason": stop_reason.value,
+            "steps": str(step),
+        }
         return last_model_response
 
     async def run(self, request: Request, body: FinanceAgentRunRequest) -> FinanceAgentVerifyResponse:
@@ -334,6 +478,7 @@ class FinanceAgent(SimpleResponsesAPIAgent):
                 tools=[],
                 parallel_tool_calls=False,
                 tool_choice="auto",
+                metadata={"stop_reason": StopReason.ERROR.value},
             )
             return FinanceAgentVerifyResponse(
                 responses_create_params=body.responses_create_params,
