+++ responses_api_agents/finance_agent/tests/test_app.py
@@ -12,14 +12,22 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for Finance Agent (responses_api_agents/finance_agent)."""
+"""Tests for Finance Agent (responses_api_agents/finance_agent).
+
+The loop is exercised under both shipped profiles, read from the configs that
+ship with the two resource servers rather than restated here, so a profile that
+stops parsing or stops setting a required field fails these tests.
+"""
 
 import asyncio
 import json
+from pathlib import Path
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
+import yaml
 from fastapi.testclient import TestClient
+from pydantic import ValidationError
 
 from nemo_gym.config_types import ModelServerRef, ResourcesServerRef
 from nemo_gym.openai_utils import (
@@ -37,13 +45,38 @@ _MODEL_SERVER = "model_server"
 _RS_SERVER = "resources_server"
 _INPUT = {"input": [{"role": "user", "content": "What was revenue?"}]}
 
+_REPO_ROOT = Path(__file__).resolve().parents[3]
+
+#: Every server instance of this loop, and the config file declaring it.
+_SHIPPED_CONFIGS = {
+    "finance_agent": _REPO_ROOT / "resources_servers/finance_sec_search/configs/finance_sec_search.yaml",
+    "finance_agent_v2": _REPO_ROOT / "resources_servers/finance_agent_v2/configs/finance_agent_v2.yaml",
+}
+
+#: Fields with no default, so each profile has to state its own policy.
+_POLICY_FIELDS = ("no_tool_call_nudge", "max_time_seconds", "abort_on_tool_error_types")
+
 
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------
 
 
-def _make_config(**overrides) -> FinanceAgentConfig:
+def _shipped_block(instance: str) -> dict:
+    config = yaml.safe_load(_SHIPPED_CONFIGS[instance].read_text())
+    return config[instance]["responses_api_agents"]["finance_agent"]
+
+
+def _shipped_policy(instance: str) -> dict:
+    block = _shipped_block(instance)
+    return {field: block[field] for field in _POLICY_FIELDS if field in block}
+
+
+_V1_POLICY = _shipped_policy("finance_agent")
+_V2_POLICY = _shipped_policy("finance_agent_v2")
+
+
+def _make_config(policy: dict | None = None, **overrides) -> FinanceAgentConfig:
     defaults = dict(
         host="0.0.0.0",
         port=8080,
@@ -52,8 +85,13 @@ def _make_config(**overrides) -> FinanceAgentConfig:
         resources_server=ResourcesServerRef(type="resources_servers", name=_RS_SERVER),
         model_server=ModelServerRef(type="responses_api_models", name=_MODEL_SERVER),
     )
+    defaults.update(_V1_POLICY if policy is None else policy)
     defaults.update(overrides)
     return FinanceAgentConfig(**defaults)
+
+
+def _v2_config(**overrides) -> FinanceAgentConfig:
+    return _make_config(policy=_V2_POLICY, **overrides)
 
 
 def _make_agent_and_client(config: FinanceAgentConfig | None = None):
@@ -150,12 +188,46 @@ def _dotjson_mock(*json_responses: dict) -> AsyncMock:
     return mock
 
 
+def _route(model_mock, rs_mock):
+    def route_post(**kwargs):
+        return model_mock if kwargs["server_name"] == _MODEL_SERVER else rs_mock
+
+    return route_post
+
+
 # ---------------------------------------------------------------------------
 # Tests: Config and Construction
 # ---------------------------------------------------------------------------
 
 
 class TestFinanceAgentConfig:
+    @pytest.mark.parametrize("instance", sorted(_SHIPPED_CONFIGS))
+    def test_shipped_profile_states_every_required_field(self, instance: str) -> None:
+        """A required field with no default is only useful if the shipped configs
+        set it; otherwise the server fails to start."""
+        block = _shipped_block(instance)
+        assert set(_POLICY_FIELDS) <= set(block), f"{instance} is missing {set(_POLICY_FIELDS) - set(block)}"
+
+    @pytest.mark.parametrize("instance", sorted(_SHIPPED_CONFIGS))
+    def test_shipped_profile_builds_a_valid_config(self, instance: str) -> None:
+        config = _make_config(policy=_shipped_policy(instance))
+        assert config.no_tool_call_nudge
+
+    @pytest.mark.parametrize("field", _POLICY_FIELDS)
+    def test_omitting_a_policy_field_is_rejected(self, field: str) -> None:
+        policy = {k: v for k, v in _V2_POLICY.items() if k != field}
+        with pytest.raises(ValidationError):
+            _make_config(policy=policy)
+
+    def test_the_two_profiles_differ_where_the_loop_branches(self) -> None:
+        """The loop has no benchmark-specific branches, so everything that makes
+        the two harnesses different has to show up here."""
+        assert _V1_POLICY["max_time_seconds"] is None
+        assert _V2_POLICY["max_time_seconds"] > 0
+        assert _V1_POLICY["abort_on_tool_error_types"] == []
+        assert _V2_POLICY["abort_on_tool_error_types"]
+        assert _V1_POLICY["no_tool_call_nudge"] != _V2_POLICY["no_tool_call_nudge"]
+
     def test_default_config(self) -> None:
         config = _make_config()
         assert config.max_steps is None
@@ -167,13 +239,19 @@ class TestFinanceAgentConfig:
     def test_custom_config(self) -> None:
         config = _make_config(
             max_steps=10,
+            max_time_seconds=60.0,
             done_tools=["submit_final_result", "abort"],
+            no_tool_call_nudge="Keep going.",
+            abort_on_tool_error_types=["RetryExhaustedError"],
             model_call_timeout=30.0,
             tool_call_timeout=60.0,
             truncate_on_overflow=True,
         )
         assert config.max_steps == 10
+        assert config.max_time_seconds == 60.0
         assert config.done_tools == ["submit_final_result", "abort"]
+        assert config.no_tool_call_nudge == "Keep going."
+        assert config.abort_on_tool_error_types == ["RetryExhaustedError"]
         assert config.model_call_timeout == 30.0
         assert config.tool_call_timeout == 60.0
         assert config.truncate_on_overflow is True
@@ -284,15 +362,63 @@ class TestTruncateOldestExchange:
 
 
 # ---------------------------------------------------------------------------
+# Tests: _aborting_error_type
+# ---------------------------------------------------------------------------
+
+
+class TestAbortingErrorType:
+    def test_matches_abort_error_in_the_servers_envelope(self) -> None:
+        """The resource server double-encodes: the error object is a JSON string
+        inside ``results``, which is the only shape a tool failure arrives in."""
+        agent, _ = _make_agent_and_client(_v2_config())
+        payload = json.dumps({"results": json.dumps({"error": "RetryExhaustedError: gave up after 5 attempts"})})
+        assert agent._aborting_error_type(payload) == "RetryExhaustedError"
+
+    def test_matches_abort_error_without_an_envelope(self) -> None:
+        """A tool call that never reached the server is rendered locally flat."""
+        agent, _ = _make_agent_and_client(_v2_config())
+        payload = json.dumps({"error": "RetryExhaustedError: gave up after 5 attempts"})
+        assert agent._aborting_error_type(payload) == "RetryExhaustedError"
+
+    def test_ignores_other_errors(self) -> None:
+        agent, _ = _make_agent_and_client(_v2_config())
+        payload = json.dumps({"results": json.dumps({"error": "ConnectionError: server unavailable"})})
+        assert agent._aborting_error_type(payload) is None
+
+    def test_ignores_successful_payload_mentioning_the_type(self) -> None:
+        """Only the error field aborts, so a tool echoing the name in its
+        results cannot kill the rollout."""
+        agent, _ = _make_agent_and_client(_v2_config())
+        payload = json.dumps({"results": "the filing mentions RetryExhaustedError: nope"})
+        assert agent._aborting_error_type(payload) is None
+
+    def test_ignores_non_json_output(self) -> None:
+        agent, _ = _make_agent_and_client(_v2_config())
+        assert agent._aborting_error_type("RetryExhaustedError: plain text") is None
+
+    def test_ignores_a_successful_reply_that_happens_to_be_json(self) -> None:
+        """A tool whose own output is a JSON object with no error field is a
+        success, however it is worded."""
+        agent, _ = _make_agent_and_client(_v2_config())
+        payload = json.dumps({"results": json.dumps({"summary": "RetryExhaustedError: appears in this filing"})})
+        assert agent._aborting_error_type(payload) is None
+
+    def test_disabled_under_the_v1_profile(self) -> None:
+        agent, _ = _make_agent_and_client()
+        payload = json.dumps({"results": json.dumps({"error": "RetryExhaustedError: gave up"})})
+        assert agent._aborting_error_type(payload) is None
+
+
+# ---------------------------------------------------------------------------
 # Tests: responses() via TestClient
 # ---------------------------------------------------------------------------
 
 
 class TestResponses:
     def test_text_only_response_injects_continue_until_max_steps(self) -> None:
-        """Text-only assistant responses must inject 'Continue.' and keep
-        looping (mirrors vals-ai/finance-agent ``_before_query`` +
-        ``_should_stop=False``) rather than terminating after one step.
+        """Text-only assistant responses must inject the nudge and keep looping
+        (mirrors vals-ai/finance-agent ``_before_query`` + ``_should_stop=False``)
+        rather than terminating after one step.
 
         The historical break-on-text behavior caused training to silently
         truncate trajectories whenever the model emitted explanatory text
@@ -317,6 +443,28 @@ class TestResponses:
             f"Continue. literal missing; got user contents: {[o['content'] for o in user_continues]}"
         )
 
+    def test_text_only_response_injects_the_v2_nudge(self) -> None:
+        """The FABv2 profile names submit_final_result instead of sending a bare
+        "Continue."; picking up the other profile's nudge is silent drift."""
+        agent, client = _make_agent_and_client(_v2_config(max_steps=3))
+        agent.server_client.post.return_value = _dotjson_mock(_text_response("Hello!"))
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+
+        nudges = [o for o in res.json()["output"] if o["type"] == "message" and o["role"] == "user"]
+        assert nudges, "no nudge injected after a text-only response"
+        assert all(o["content"] == _V2_POLICY["no_tool_call_nudge"] for o in nudges)
+
+    def test_nudge_is_configurable(self) -> None:
+        agent, client = _make_agent_and_client(_make_config(max_steps=1, no_tool_call_nudge="Keep going."))
+        agent.server_client.post.return_value = _dotjson_mock(_text_response("Hello!"))
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        nudges = [o for o in res.json()["output"] if o["type"] == "message" and o["role"] == "user"]
+        assert [o["content"] for o in nudges] == ["Keep going."]
+
     def test_continue_injection_stops_at_submit_final_result(self) -> None:
         """Continue.-loop must yield as soon as a done-tool fires -- otherwise
         the agent could keep looping past a legitimate terminal tool call.
@@ -331,12 +479,7 @@ class TestResponses:
         model_mock = _dotjson_mock(*text_then_submit_responses)
         rs_mock = _dotjson_mock({"status": "ok"})
 
-        def route_post(**kwargs):
-            if kwargs["server_name"] == _MODEL_SERVER:
-                return model_mock
-            return rs_mock
-
-        agent.server_client.post = AsyncMock(side_effect=route_post)
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
 
         res = client.post("/v1/responses", json=_INPUT)
         assert res.status_code == 200
@@ -359,12 +502,7 @@ class TestResponses:
         model_mock = _dotjson_mock(tool_call_data, text_data)
         rs_mock = _dotjson_mock(tool_result)
 
-        def route_post(**kwargs):
-            if kwargs["server_name"] == _MODEL_SERVER:
-                return model_mock
-            return rs_mock
-
-        agent.server_client.post = AsyncMock(side_effect=route_post)
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
 
         res = client.post("/v1/responses", json=_INPUT)
         assert res.status_code == 200
@@ -388,12 +526,7 @@ class TestResponses:
         model_mock = _dotjson_mock(tool_call_data)
         rs_mock = _dotjson_mock(tool_result)
 
-        def route_post(**kwargs):
-            if kwargs["server_name"] == _MODEL_SERVER:
-                return model_mock
-            return rs_mock
-
-        agent.server_client.post = AsyncMock(side_effect=route_post)
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
 
         res = client.post("/v1/responses", json=_INPUT)
         assert res.status_code == 200
@@ -412,18 +545,143 @@ class TestResponses:
         model_mock = _dotjson_mock(tool_call_data, tool_call_data)
         rs_mock = _dotjson_mock(tool_result)
 
-        def route_post(**kwargs):
-            if kwargs["server_name"] == _MODEL_SERVER:
-                return model_mock
-            return rs_mock
-
-        agent.server_client.post = AsyncMock(side_effect=route_post)
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
 
         res = client.post("/v1/responses", json=_INPUT)
         assert res.status_code == 200
         output = res.json()["output"]
         fn_calls = [o for o in output if o["type"] == "function_call"]
         assert len(fn_calls) == 2
+
+    def test_exhausted_time_budget_prevents_any_model_call(self) -> None:
+        """The budget is checked before the turn's query, so an already-expired
+        budget must not pay for one more call."""
+        agent, client = _make_agent_and_client(_v2_config(max_time_seconds=0))
+        agent.server_client.post = AsyncMock()
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        assert res.json()["id"] == "error"
+        assert agent.server_client.post.call_count == 0
+
+    def test_time_budget_stops_loop_mid_run(self) -> None:
+        agent, client = _make_agent_and_client(_v2_config(max_steps=5, max_time_seconds=0.05))
+
+        tool_call = _tool_call_response("edgar_search", json.dumps({"search_query": "AAPL"}))
+        rs_mock = _dotjson_mock({"results": "data"})
+
+        async def route_post(**kwargs):
+            if kwargs["server_name"] == _MODEL_SERVER:
+                await asyncio.sleep(0.06)
+                return _dotjson_mock(tool_call)
+            return rs_mock
+
+        agent.server_client.post = AsyncMock(side_effect=route_post)
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        model_calls = [c for c in agent.server_client.post.call_args_list if c.kwargs["server_name"] == _MODEL_SERVER]
+        assert len(model_calls) == 1, "budget should have expired before the second turn"
+
+    def test_no_time_budget_under_the_v1_profile(self) -> None:
+        """max_time_seconds: null must leave the loop turn-bounded, with no clock
+        check able to cut a training rollout short."""
+        agent, client = _make_agent_and_client(_make_config(max_steps=2))
+
+        tool_call = _tool_call_response("sec_filing_search", json.dumps({"ticker": "AAPL"}))
+        model_mock = _dotjson_mock(tool_call, tool_call)
+
+        async def route_post(**kwargs):
+            if kwargs["server_name"] == _MODEL_SERVER:
+                await asyncio.sleep(0.02)
+                return model_mock
+            return _dotjson_mock({"result": "data"})
+
+        agent.server_client.post = AsyncMock(side_effect=route_post)
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        assert res.json()["metadata"]["stop_reason"] == "max_turns"
+
+    def test_stop_reason_recorded_for_a_clean_submission(self) -> None:
+        agent, client = _make_agent_and_client()
+
+        model_mock = _dotjson_mock(_tool_call_response("submit_final_result", json.dumps({"final_result": "$100B"})))
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, _dotjson_mock({"status": "ok"})))
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        assert res.json()["metadata"]["stop_reason"] == "done_tool"
+        assert res.json()["metadata"]["steps"] == "1"
+
+    @pytest.mark.parametrize(
+        "config, expected",
+        [
+            (dict(max_steps=2), "max_turns"),
+            (dict(max_steps=5, max_time_seconds=0.05), "max_time"),
+        ],
+    )
+    def test_stop_reason_distinguishes_truncated_rollouts(self, config, expected) -> None:
+        """A trajectory cut short scores like a confident miss under dealbreaker
+        gating, so the results file has to say which one it was."""
+        agent, client = _make_agent_and_client(_v2_config(**config))
+        tool_call = _tool_call_response("edgar_search", json.dumps({"search_query": "AAPL"}))
+        rs_mock = _dotjson_mock({"results": "data"})
+
+        async def route_post(**kwargs):
+            if kwargs["server_name"] == _MODEL_SERVER:
+                if config.get("max_time_seconds"):
+                    await asyncio.sleep(0.06)
+                return _dotjson_mock(tool_call)
+            return rs_mock
+
+        agent.server_client.post = AsyncMock(side_effect=route_post)
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        assert res.json()["metadata"]["stop_reason"] == expected
+
+    def test_abort_error_terminates_rollout(self) -> None:
+        """The FABv2 profile aborts on retry exhaustion rather than letting the
+        model continue without the data."""
+        agent, client = _make_agent_and_client(_v2_config(max_steps=5))
+
+        tool_call = _tool_call_response("price_history", json.dumps({"ticker": "AAPL"}))
+        model_mock = _dotjson_mock(tool_call, tool_call)
+        rs_mock = _dotjson_mock({"results": json.dumps({"error": "RetryExhaustedError: gave up after 5 attempts"})})
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        model_calls = [c for c in agent.server_client.post.call_args_list if c.kwargs["server_name"] == _MODEL_SERVER]
+        assert len(model_calls) == 1, "rollout should abort instead of starting a second turn"
+
+    def test_ordinary_tool_error_does_not_abort(self) -> None:
+        agent, client = _make_agent_and_client(_v2_config(max_steps=2))
+
+        tool_call = _tool_call_response("price_history", json.dumps({"ticker": "AAPL"}))
+        model_mock = _dotjson_mock(tool_call, tool_call)
+        rs_mock = _dotjson_mock({"results": json.dumps({"error": "ConnectionError: server unavailable"})})
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        model_calls = [c for c in agent.server_client.post.call_args_list if c.kwargs["server_name"] == _MODEL_SERVER]
+        assert len(model_calls) == 2, "a plain tool failure is fed back to the model"
+
+    def test_abort_error_is_fed_back_under_the_v1_profile(self) -> None:
+        """With no abort types configured, the same error keeps the loop going."""
+        agent, client = _make_agent_and_client(_make_config(max_steps=2))
+
+        tool_call = _tool_call_response("sec_filing_search", json.dumps({"ticker": "AAPL"}))
+        model_mock = _dotjson_mock(tool_call, tool_call)
+        rs_mock = _dotjson_mock({"results": json.dumps({"error": "RetryExhaustedError: gave up after 5 attempts"})})
+        agent.server_client.post = AsyncMock(side_effect=_route(model_mock, rs_mock))
+
+        res = client.post("/v1/responses", json=_INPUT)
+        assert res.status_code == 200
+        model_calls = [c for c in agent.server_client.post.call_args_list if c.kwargs["server_name"] == _MODEL_SERVER]
+        assert len(model_calls) == 2
 
     def test_model_call_timeout(self) -> None:
         """Model call timeout → loop terminates gracefully."""
