--- responses_api_agents/finance_agent_v2/requirements.txt
@@ -1,20 +1,0 @@
-# Finance Agent v2 (FABv2) Agent Dependencies
-#
-# Unlike responses_api_agents/finance_agent (nemo-gym only), this agent depends
-# on the upstream package so that every parity-relevant value can be imported
-# instead of copied: the done-tool name, the 1h time budget, the valid tool
-# names, and the abort-worthy exception type. The pins are exact and must stay
-# byte-identical to resources_servers/finance_agent_v2/requirements.txt —
-# tests/test_app.py asserts they match, so the agent and the tools it drives can
-# never be built from two different upstream commits.
--e nemo-gym[dev] @ ../../
-
-# Vals model-library (transitive via finance-agent; pinned explicitly so both
-# components resolve the same version). Its metadata declares
-# openai[aiohttp]>=2.28.0, which cannot resolve against the vLLM-compatible openai
-# nemo-gym pins; `overrides.txt` drops that floor.
-model-library==0.1.25
-
-# Vals finance-agent-v2 (provides finance_agent.get_agent / .tools / .exceptions).
-# No upstream tags exist; pinned to the HEAD commit at integration time.
-finance-agent @ git+https://github.com/vals-ai/finance-agent-v2.git@bb460b524472ae0955abdf243bbbac06958dc8c9
